Add files via upload
Cats and Dogs

Create a class Dog with a name and an age. The class should also have a bark function that prints 'name of dog: Woof'
Create a class Cat with a name and an age. The class should also have a meow function that prints 'name of cat: Meow'

Create a list of five dogs and a list of five cats with random ages
Use a loop to iterate through the list of cats and make each cat meow, also use a nested loop to make each dog bark if they are older than the meowing cat

Print everything to the console

Example 1:

Cat 2 is 5 years old and meows

Dogs 1 and 2 are 3 years old so they stay quiet, while dogs 3, 4, and 5 are older than the cat so they bark

Example 2:

In this example, the cats are Lili, Meep, Patchy, Furball, and Snowball

Each dog that barks is older than the cat that is meowing.